Update patch set 1
Patch Set 1:

> Not sure that _this_ is causing CI issues, because python-* will
 > still pull python2-* which has latest NVR, issue is in puppet not
 > supporting allow_virtual.
 > But this change is fine.

In p-o-i jobs python-setuptools is installed before enabling delorean-*-testing so python-setuptools-0.9.8 in base repo is installed. After that nothing is asking to update it until stevedore is installed which installs python2-setuptools since it was added as a dependency. This is what i've seen in job logs.

Anyway, as soon as any other package will require setuptools > 0.9.8 (the version in base), we'll have the issue back.

Patch-set: 1